Create Sort File Names
Many thanks to @sp4ghet, who helped out immensely.

This takes advantage of the default file naming system of the in-studio recording device to sort DJ recordings based of the time and date of recording. File names are formatted MDDYYYYHHMMSS, so the code has to be manually updated every week to keep up with the changing week.